Add files via upload
Social networking has become a salient part of our information media. It affects the beliefs, values, an attitude of people, as well as their behaviors. Therefore, converting information media content into a meaningful information, key concepts and themes are vital for generating knowledge and formulating strategies. Large-scale analysis of information media content allows for real-time discovery of a macro-scale token pattern in public opinion and sentiment. In this research, we analyze a collection of 26 thousand tweets generated by more than 11 thousand users around the globe in a real time.
The analysis of news content is traditionally based on the coding of data by human coders. Now measuring the current public mood is a challenging task on a huge collection of data in real time. Public media such as Twitter or Facebook can easily become a valuable source of information about the public due to the fact that people use them to express their feelings in public media. It is feasible to capture the public mood by monitoring the stream of Twitter data. Our main focus is to track four moods which are Fear, Joy, Happy and Sadness. For each mood, we track a long list of associated words and we count the frequencies that these words appear in the tweet. We describe how the analysis of Twitter content can reveal mood changes in entire populations, most crucial this survey aims to demonstrate some of the steps that can be automated, allowing researchers to access a macroscopic token pattern that would be otherwise out of reach.